# src/gittest/swagger.yaml
openapi: 3.0.0
info:
  title: Marketplace API
  version: 1.0.0
servers: [
    {
      "url": "http://localhost:5000"
    }
  ]
paths:
...
  /api/v2/payments:
    get:
      tags:
        - payments
      summary: Get all payments
      security:
        - jwt: []
      responses:
        '200':
          description: List of found payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Payments not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
  /api/v2/payments/{payment_id}:
    delete:
      tags:
        - payments
      summary: Delete a payment
      security:
        - jwt: []
      parameters:
        - name: payment_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Payment deleted
        '401':
          description: Unauthorized
        '404':
          description: Payment not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
    patch:
      tags:
        - payments
      summary: Update a payment
      security:
        - jwt: []
      parameters:
        - name: payment_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                all_price:
                  type: number
                state:
                  type: string
                user_id:
                  type: integer
      responses:
        '200':
          description: User updated 
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error
...
  /api/v2/orders:
    get:
      tags:
        - orders
      summary: Get all orders
      security:
        - jwt: []
      responses:
        '200':
          description: List of found orders
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Orders not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
  /api/v2/orders/{order_id}:
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the order
    get:
      tags:
        - orders
      summary: Get order by ID
      security:
        - jwt: []
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
        '404':
          description: Order not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
    patch:
      tags:
        - orders
      summary: Update an order
      security:
        - jwt: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                all_price:
                  type: number
                address:
                  type: string
                user_id:
                  type: integer
      responses:
        '200':
          description: Order updated
        '401':
          description: Unauthorized
        '404':
          description: Order not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      type: object
      properties:
        payment_id:
          type: integer
        all_price:
          type: number
        state:
          type: string
        user_id:
          type: integer
      required:
        - payment_id
        - all_price
        - state
        - user_id
    Order:
      type: object
      properties:
        order_id:
          type: integer
        all_price:
          type: number
        address:
          type: string
        user_id:
          type: integer
      required:
        - order_id
        - all_price
        - address
        - user_id
...